MRD-2987 Move SentenceSequence to new package
